linux-local: Fix build with local defconfig with absoluth path
Commit 8841655888 ("linux-local: Add variable for local defconfig
with absoluth path") has an uncovered scenario: if the file used
as defconfig does not include in his name the word "defconfig",
the build will fail because on do_kernel_metadata bitbake is doing
a check if SRC_URI contains this word, and if not it goes to other
methods of creating the configuration file. If none works, it fails:

ERROR: Could not locate BSP definition for intel-axxia-grr/standard
       and no defconfig was provided

For instance, a file .defconfig-for-6.13 works fine, but one like
.config-for-6.13 not.

With this patch, the file can have any name with no limitations.

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>