cpupower: Fix installed-vs-shipped QA error for kernel v6.16
Starting with commit 9c70b779ad ("cpupower: add a systemd service
to run cpupower") from kernel v6.16, a new systemd service is created
which is not shipped in any package.

Fix this by including it in the main package for systemd and remove
it for sysvinit.

ERROR: cpupower-1.0-r0 do_package: QA Issue: cpupower:
       Files/directories were installed but not shipped in any package:
  /usr/lib64/systemd
  /usr/lib64/systemd/system
  /usr/lib64/systemd/system/cpupower.service

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>