linux-local: Set KCONFIG_MODE to alldefconfig
When the "*defconfig*" file used for linux-local recipe is not
a full .config containing all the kernel option, but a defconfig
from kernel sources containing only the different options
compared to Kconfig files, we need to decide how the missing
options will be filled. If KCONFIG_MODE is not set, the default
value is "allnoconfig" (new options disabled) and will create a
very different .config. With "alldefconfig" the new options gets
the default value and will create a final .config similar with
the one used to create the defconfig whith "make savedefconfig".

B0330 - Build Fails when using linux-local

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>