linux-local: Add ability to build from a remote repository
The main purpose of linux-local recipe is to build linux from a
local repository where the development is done. This new feature
allow the recipe to fetch the linux sources from a remote
repository. If LOCAL_KERNEL_PATH points to a local path will
clone from there, if the value starts with http*, will fetch
source from that remote repository using https protocol.

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>