fio: Add initial recipe for 3.41 version
Fio (flexible I/O tester) is a tool that will spawn a number of threads or
processes doing a particular type of I/O action as specified by the user.
The typical use of fio is to write a job file matching the I/O load one
wants to simulate.

Add fio to -dev and -run images.

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>