linux-korg-netnext: Use defconfig from KORG_NETNEXT_EXTRA_PATH
Allow using a defconfig from the extra path instead of kernel fragments.
A defconfig can also be specified via KORG_NETNEXT_CONFIG. See bellow:

KORG_NETNEXT_EXTRA_PATH variable can be used in local.conf to specify
a path containing patches, kernel fragments, or a defconfig. All of
them will be used to patch the kernel sources and generate the final
.config. If the variable is not set, the final .config will be generated
from fragments. If it is set, BitBake will look for a *defconfig file
the file name must end with defconfig) in the provided path.

KORG_NETNEXT_CONFIG variable can be set in local.conf to specify the full
path to a defconfig file (the file name must end with defconfig).

Signed-off-by: Daniel Dragomir <daniel.dragomir@windriver.com>